refactor(viewIn, viewOut): view 관련 클래스 함수화
- View에 관한 것을 Class로 관리했는데, 함수화 하였습니다.
- View 내부 매서드간에 상태를 공유하지 않고 각 기능이 독립적이라 판단하여서 리팩토링 하였습니다.
